+++ android/src/main/java/com/github/drydart/flutter_android/SensorManagerHandler.java
@@ -5,6 +5,7 @@ package com.github.drydart.flutter_android;
 import android.content.Context;
 import android.hardware.Sensor;
 import android.hardware.SensorManager;
+import io.flutter.plugin.common.EventChannel;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel.Result;
 import io.flutter.plugin.common.PluginRegistry.Registrar;
@@ -16,7 +17,8 @@ import java.util.Map;
 class SensorManagerHandler extends FlutterMethodCallHandler {
   static final String CHANNEL = "flutter_android/SensorManager";
 
-  final Map<Integer, Sensor> sensors = new HashMap<>();
+  private final Map<Integer, Sensor> sensors = new HashMap<>();
+  private final Map<Integer, EventChannel> channels = new HashMap<>();
 
   SensorManagerHandler(final Registrar registrar) {
     super(registrar);
@@ -33,8 +35,9 @@ class SensorManagerHandler extends FlutterMethodCallHandler {
 
       case "getDefaultSensor": {
         final int sensorType = getRequiredArgument(call, "type");
-        if (this.sensors.containsKey(sensorType)) {
-          result.success(sensorType);
+        final int sensorKey = sensorType; // for now
+        if (this.sensors.containsKey(sensorKey)) {
+          result.success(sensorKey);
           break;
         }
         final Context context = registrar.context();
@@ -44,8 +47,38 @@ class SensorManagerHandler extends FlutterMethodCallHandler {
           result.success(null);
           break;
         }
-        this.sensors.put(sensorType, sensor);
+        this.sensors.put(sensorKey, sensor);
         result.success(sensorType);
+        break;
+      }
+
+      /* Instance methods */
+
+      case "registerListener": {
+        final int sensorKey = getRequiredArgument(call, "key");
+        if (this.channels.containsKey(sensorKey)) {
+          result.success(true); // already registered
+          break;
+        }
+        if (!this.sensors.containsKey(sensorKey)) {
+          result.success(false); // invalid sensor key
+          break;
+        }
+        final int samplingPeriodUs = getOptionalArgument(call, "samplingPeriodUs", SensorManager.SENSOR_DELAY_NORMAL);
+        final int maxReportLatencyUs = getOptionalArgument(call, "maxReportLatencyUs", 0);
+        final Context context = registrar.context();
+        final SensorManager sensorManager = (SensorManager)context.getSystemService(Context.SENSOR_SERVICE);
+        final Sensor sensor = this.sensors.get(sensorKey);
+        final String channelID = String.valueOf(sensorKey);
+        final EventChannel channel = new EventChannel(registrar.messenger(), String.format("%s/%d", CHANNEL, sensorKey));
+        channel.setStreamHandler(new SensorEventStream(sensorManager, sensor, samplingPeriodUs, maxReportLatencyUs));
+        this.channels.put(sensorKey, channel);
+        result.success(channelID);
+        break;
+      }
+
+      case "unregisterListener": {
+        // TODO
         break;
       }